refactor: move components folder to src
