+++ src/components/Slide.tsx
